fast-api-backend/requirements.txt
adal==1.2.7
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
azure-common==1.1.28
azure-core==1.35.0
azure-graphrbac==0.61.2
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==13.0.0
azure-mgmt-core==1.6.0
azure-mgmt-keyvault==11.0.0
azure-mgmt-network==29.0.0
azure-mgmt-resource==24.0.0
azure-mgmt-storage==23.0.0
azureml-core==1.60.0.post1
babel==2.17.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.3.0
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
boto==2.49.0
boto3==1.40.25
botocore==1.40.25
cachetools==5.5.2
catboost==1.2.8
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
comm==0.2.3
contextlib2==21.6.0
contourpy==1.3.2
cramjam==2.11.0
cryptography==45.0.7
cycler==0.12.1
dash==3.2.0
databricks-sdk==0.65.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
dnspython==2.8.0
docker==7.1.0
durationpy==0.10
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastjsonschema==2.21.2
fastparquet==2024.11.0
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.2
fqdn==1.5.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaggle==1.7.4.5
kiwisolver==1.4.9
knack==0.12.0
kubernetes==33.1.0
lark==1.2.2
lightgbm==4.6.0
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.16
mistune==3.1.4
mlflow==2.20.0
mlflow-skinny==2.20.0
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
msrestazure==0.6.4.post1
narwhals==2.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.2.6
nvidia-nccl-cu12==2.28.3
oauthlib==3.3.1
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
overrides==7.7.0
packaging==24.2
pandas==2.3.2
pandocfilters==1.5.1
paramiko==3.5.1
parquet==1.3.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
pipenv==2025.0.4
pkginfo==1.12.1.2
platformdirs==4.4.0
plotly==6.3.0
pluggy==1.6.0
ply==3.11
prometheus_client==0.22.1
prometheus_flask_exporter==0.23.2
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pysftp==0.2.9
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
requests-auth-aws-sigv4==0.7
requests-oauthlib==2.0.0
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.3
sentry-sdk==2.38.0
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.3
tabulate==0.9.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
thriftpy2==0.5.3
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.17.4
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
xgboost==3.0.5
zipp==3.23.0
